+++ src/bitvavo_client/endpoints/public.py
@@ -4,11 +4,12 @@ from __future__ import annotations
 
 from typing import TYPE_CHECKING, Any, Literal, TypeVar
 
-from returns.result import Failure, Success
+from returns.result import Result
 
 from bitvavo_client.adapters.returns_adapter import BitvavoError
 from bitvavo_client.core import public_models
 from bitvavo_client.core.model_preferences import ModelPreference
+from bitvavo_client.endpoints.base import BaseAPI
 from bitvavo_client.endpoints.common import create_postfix
 from bitvavo_client.schemas.public_schemas import DEFAULT_SCHEMAS
 
@@ -24,31 +25,33 @@ MAX_TIMESTAMP_VALUE = 8640000000000000  # Maximum allowed timestamp value
 MAX_BOOK_DEPTH = 1000  # Maximum depth for order book
 MAX_BOOK_REPORT_DEPTH = 1000  # Maximum depth for order book report
 
-# DataFrames preference to library mapping
-_DATAFRAME_LIBRARY_MAP = {
-    ModelPreference.POLARS: ("polars", "pl.DataFrame"),
-    ModelPreference.PANDAS: ("pandas", "pd.DataFrame"),
-    ModelPreference.PYARROW: ("pyarrow", "pa.Table"),
-    ModelPreference.DASK: ("dask", "dd.DataFrame"),
-    ModelPreference.MODIN: ("modin", "mpd.DataFrame"),
-    ModelPreference.CUDF: ("cudf", "cudf.DataFrame"),
-    ModelPreference.IBIS: ("ibis", "ibis.Table"),
-}
-
 if TYPE_CHECKING:  # pragma: no cover
     from collections.abc import Mapping
-
     import httpx
-    from returns.result import Result
-
     from bitvavo_client.core.types import AnyDict
     from bitvavo_client.transport.http import HTTPClient
 
 T = TypeVar("T")
 
 
-class PublicAPI:
+class PublicAPI(BaseAPI):
     """Handles all public Bitvavo API endpoints."""
+
+    _endpoint_models = {
+        "time": public_models.ServerTime,
+        "markets": public_models.Markets,
+        "assets": public_models.Assets,
+        "book": public_models.OrderBook,
+        "trades": public_models.Trades,
+        "candles": public_models.Candles,
+        "ticker_price": public_models.TickerPrices,
+        "ticker_book": public_models.TickerBooks,
+        "ticker_24h": public_models.Ticker24hs,
+        "report_book": public_models.OrderBookReport,
+        "report_trades": public_models.TradesReport,
+    }
+
+    _default_schemas = DEFAULT_SCHEMAS
 
     def __init__(
         self,
@@ -57,209 +60,10 @@ class PublicAPI:
         preferred_model: ModelPreference | str | None = None,
         default_schema: Mapping[str, object] | None = None,
     ) -> None:
-        """Initialize public API handler.
-
-        Args:
-            http_client: HTTP client for making requests
-            preferred_model: Preferred model format for responses
-            default_schema: Default schema for DataFrame conversion
-        """
-        self.http: HTTPClient = http_client
-
-        # Handle preferred_model parameter - try to convert strings to ModelPreference,
-        # but allow arbitrary strings to pass through for custom handling
-        if preferred_model is None:
-            self.preferred_model = None
-        elif isinstance(preferred_model, ModelPreference):
-            self.preferred_model = preferred_model
-        elif isinstance(preferred_model, str):
-            try:
-                self.preferred_model = ModelPreference(preferred_model)
-            except ValueError:
-                # If string doesn't match a valid ModelPreference, store as-is
-                self.preferred_model = preferred_model
-        else:
-            self.preferred_model = preferred_model
-
-        # If using DATAFRAME preference without a default schema, we could provide sensible defaults
-        # But keep it explicit for now - users can import and use schemas as needed
-        self.default_schema = default_schema
-
-    def _get_effective_model(
-        self,
-        endpoint_type: str,
-        model: type[T] | Any | None,
-        schema: Mapping[str, object] | None,
-    ) -> tuple[type[T] | Any | None, Mapping[str, object] | None]:
-        """Get the effective model and schema to use for a request.
-
-        Args:
-            endpoint_type: Type of endpoint (e.g., 'time', 'markets', 'assets')
-            model: Model explicitly passed to method (overrides preference)
-            schema: Schema explicitly passed to method
-
-        Returns:
-            Tuple of (effective_model, effective_schema)
-        """
-        # If model is explicitly provided, use it
-        if model is not None:
-            return model, schema
-
-        # If no preferred model is set, return Any (raw response)
-        if self.preferred_model is None:
-            return Any, schema
-
-        # Apply preference based on enum value
-        if self.preferred_model == ModelPreference.RAW:
-            return Any, schema
-
-        # Handle all DataFrame preferences
-        if self.preferred_model in _DATAFRAME_LIBRARY_MAP:
-            # Use the provided schema, fallback to instance default, then to endpoint-specific default
-            effective_schema = schema or self.default_schema or DEFAULT_SCHEMAS.get(endpoint_type)
-            # Return the preference itself, not a specific DataFrame class
-            return self.preferred_model, effective_schema
-
-        if self.preferred_model == ModelPreference.PYDANTIC:
-            # Map endpoint types to appropriate Pydantic models
-            endpoint_model_map = {
-                "time": public_models.ServerTime,
-                "markets": public_models.Markets,
-                "assets": public_models.Assets,
-                "book": public_models.OrderBook,
-                "trades": public_models.Trades,
-                "candles": public_models.Candles,
-                "ticker_price": public_models.TickerPrices,
-                "ticker_book": public_models.TickerBooks,
-                "ticker_24h": public_models.Ticker24hs,
-                "report_book": public_models.OrderBookReport,
-                "report_trades": public_models.TradesReport,
-            }
-            return endpoint_model_map.get(endpoint_type, dict), schema
-
-        # Default case (AUTO or unknown)
-        return None, schema
-
-    def _convert_raw_result(
-        self,
-        raw_result: Result[Any, BitvavoError | httpx.HTTPError],
-        endpoint_type: str,
-        model: type[T] | Any | None,
-        schema: Mapping[str, object] | None,
-    ) -> Result[Any, BitvavoError | httpx.HTTPError]:
-        """Convert raw API result to the desired model format.
-
-        Args:
-            raw_result: Raw result from HTTP client
-            endpoint_type: Type of endpoint (e.g., 'time', 'markets', 'assets')
-            model: Model explicitly passed to method (overrides preference)
-            schema: Schema explicitly passed to method
-
-        Returns:
-            Result with converted data or original error
-        """
-        # If the raw result is an error, return it as-is
-        if isinstance(raw_result, Failure):
-            return raw_result
-
-        # Get the effective model and schema to use
-        effective_model, effective_schema = self._get_effective_model(endpoint_type, model, schema)
-
-        # If no conversion needed (raw data requested), return as-is
-        if effective_model is Any or effective_model is None or effective_model == ModelPreference.RAW:
-            return raw_result
-
-        # Extract the raw data
-        raw_data = raw_result.unwrap()
-
-        # Perform conversion
-        try:
-            # Handle DataFrame preferences specially
-            if isinstance(effective_model, ModelPreference) and effective_model in _DATAFRAME_LIBRARY_MAP:
-                parsed = self._create_dataframe(raw_data, effective_model, effective_schema)
-            elif hasattr(effective_model, "model_validate"):
-                # Pydantic model
-                parsed = effective_model.model_validate(raw_data)  # type: ignore[misc]
-            else:
-                # Simple constructor call - this handles dict and other simple types
-                parsed = effective_model(raw_data)  # type: ignore[misc]
-
-            return Success(parsed)
-        except (ValueError, TypeError, AttributeError) as exc:
-            # If conversion fails, return a structured error
-            error = BitvavoError(
-                http_status=500,
-                error_code=-1,
-                reason="Model conversion failed",
-                message=str(exc),
-                raw=raw_data if isinstance(raw_data, dict) else {"raw": raw_data},
-            )
-            return Failure(error)
-
-    def _create_dataframe(
-        self,
-        data: Any,
-        preference: ModelPreference,
-        schema: Mapping[str, object] | None,
-    ) -> Any:
-        """Create a DataFrame from raw data using the specified preference.
-
-        Args:
-            data: Raw data to convert
-            preference: DataFrame preference (POLARS, PANDAS, etc.)
-            schema: Schema for DataFrame conversion
-
-        Returns:
-            DataFrame instance
-
-        Raises:
-            ImportError: If the required DataFrame library is not available
-            ValueError: If DataFrame creation fails
-        """
-        if preference == ModelPreference.POLARS:
-            import polars as pl
-
-            return self._create_polars_dataframe(data, schema, pl)
-
-        if preference == ModelPreference.PANDAS:
-            import pandas as pd
-
-            return pd.DataFrame(data)
-
-        if preference == ModelPreference.PYARROW:
-            import pyarrow as pa
-
-            return pa.Table.from_pylist(data if isinstance(data, list) else [data])
-
-        # For other DataFrame types, try basic conversion
-        msg = f"DataFrame preference {preference} not yet fully implemented"
-        raise NotImplementedError(msg)
-
-    def _create_polars_dataframe(
-        self,
-        data: Any,
-        schema: Mapping[str, object] | None,
-        pl: Any,  # polars module
-    ) -> Any:
-        """Create a Polars DataFrame with proper schema handling.
-
-        Args:
-            data: Raw data to convert
-            schema: Schema for DataFrame conversion
-            pl: Polars module
-
-        Returns:
-            Polars DataFrame
-        """
-        if schema is None:
-            return pl.DataFrame(data)
-
-        # For Polars, we need to handle the schema with strict=False for compatibility
-        try:
-            return pl.DataFrame(data, schema=schema, strict=False)
-        except Exception:
-            # Fallback to basic DataFrame creation if schema fails
-            return pl.DataFrame(data)
+        """Initialize public API handler."""
+        super().__init__(
+            http_client, preferred_model=preferred_model, default_schema=default_schema
+        )
 
     def time(
         self,
